refact: move vendors resolver into Host object
