+++ server/graphql/v2/interface/Account.ts
@@ -1,12 +1,10 @@
 import { GraphQLBoolean, GraphQLInt, GraphQLInterfaceType, GraphQLList, GraphQLNonNull, GraphQLString } from 'graphql';
 import { GraphQLDateTime, GraphQLJSON } from 'graphql-scalars';
 import { assign, get, invert, isEmpty, isNull, merge, omitBy } from 'lodash';
-import Sequelize, { Order } from 'sequelize';
+import { Order } from 'sequelize';
 
 import { CollectiveType } from '../../../constants/collectives';
 import FEATURE from '../../../constants/feature';
-import POLICIES from '../../../constants/policies';
-import { getPolicy } from '../../../lib/policies';
 import { buildSearchConditions } from '../../../lib/search';
 import { getCollectiveFeed } from '../../../lib/timeline';
 import { canSeeLegalName } from '../../../lib/user-permissions';
@@ -57,7 +55,6 @@ import { GraphQLPolicies } from '../object/Policies';
 import { GraphQLSocialLink } from '../object/SocialLink';
 import { GraphQLTagStats } from '../object/TagStats';
 import { GraphQLTransferWise } from '../object/TransferWise';
-import { GraphQLVendor } from '../object/Vendor';
 import { OrdersCollectionArgs, OrdersCollectionResolver } from '../query/collection/OrdersCollectionQuery';
 import {
   TransactionsCollectionArgs,
@@ -724,24 +721,6 @@ const accountFieldsDefinition = () => ({
       }
     },
   },
-  vendors: {
-    type: new GraphQLList(GraphQLVendor),
-    description: 'Get the list of vendors for this account',
-    args: {
-      forAccount: {
-        type: GraphQLAccountReferenceInput,
-        description: 'Rank vendors based on their relationship with this account',
-      },
-      isArchived: {
-        type: GraphQLBoolean,
-        description: 'Filter on archived vendors',
-      },
-      searchTerm: {
-        type: GraphQLString,
-        description: 'Search vendors related to this term based on name, description, tags, slug, and location',
-      },
-    },
-  },
 });
 
 export const GraphQLAccount = new GraphQLInterfaceType({
@@ -1056,78 +1035,6 @@ export const AccountFields = {
     type: new GraphQLNonNull(GraphQLAccountPermissions),
     description: 'Logged-in user permissions on an account',
     resolve: collective => collective, // Individual resolvers in `AccountPermissions`
-  },
-  vendors: {
-    type: new GraphQLList(GraphQLVendor),
-    description: 'Returns a list of vendors that works with this host',
-    args: {
-      forAccount: {
-        type: GraphQLAccountReferenceInput,
-        description: 'Rank vendors based on their relationship with this account',
-      },
-      isArchived: {
-        type: GraphQLBoolean,
-        description: 'Filter on archived vendors',
-      },
-      searchTerm: {
-        type: GraphQLString,
-        description: 'Search vendors related to this term based on name, description, tags, slug, and location',
-      },
-      limit: {
-        type: GraphQLInt,
-        description: 'Limit the number of vendors returned',
-      },
-    },
-    async resolve(account, args, req) {
-      const where = {
-        ParentCollectiveId: account.id,
-        type: CollectiveType.VENDOR,
-        deactivatedAt: { [args.isArchived ? Op.not : Op.is]: null },
-      };
-
-      const publicVendorPolicy = await getPolicy(account, POLICIES.EXPENSE_PUBLIC_VENDORS);
-      const isAdmin = req.remoteUser.isAdminOfCollective(account);
-      if (!publicVendorPolicy && !isAdmin) {
-        return [];
-      }
-
-      const searchTermConditions =
-        args?.searchTerm &&
-        buildSearchConditions(args.searchTerm, {
-          idFields: ['id'],
-          slugFields: ['slug'],
-          textFields: ['name', 'description', 'longDescription'],
-          stringArrayFields: ['tags'],
-          stringArrayTransformFn: str => str.toLowerCase(), // collective tags are stored lowercase
-        });
-      if (searchTermConditions?.length) {
-        where[Op.or] = searchTermConditions;
-      }
-
-      const findArgs = { where, limit: args.limit };
-      if (args?.forAccount) {
-        const account = await fetchAccountWithReference(args.forAccount);
-        findArgs['attributes'] = {
-          include: [
-            [
-              Sequelize.literal(`(
-            SELECT COUNT(*) FROM "Expenses" WHERE "deletedAt" IS NULL AND "status" = 'PAID' AND "CollectiveId" = ${account.id} AND "FromCollectiveId" = "Collective"."id"
-          )`),
-              'expenseCount',
-            ],
-          ],
-        };
-        findArgs['order'] = [[Sequelize.literal('"expenseCount"'), 'DESC']];
-      }
-
-      const vendors = await models.Collective.findAll(findArgs);
-
-      if (args?.forAccount && !isAdmin) {
-        return vendors.filter(v => v.dataValues['expenseCount'] > 0);
-      } else {
-        return vendors;
-      }
-    },
   },
 };
