+++ server/graphql/v2/object/Vendor.ts
@@ -3,7 +3,6 @@ import { GraphQLBoolean, GraphQLObjectType, GraphQLString } from 'graphql';
 import { CollectiveType } from '../../../constants/collectives';
 import { AccountFields, GraphQLAccount } from '../interface/Account';
 import { AccountWithContributionsFields, GraphQLAccountWithContributions } from '../interface/AccountWithContributions';
-import { AccountWithHostFields, GraphQLAccountWithHost } from '../interface/AccountWithHost';
 
 const GraphQLVendorContact = new GraphQLObjectType({
   name: 'VendorContact',
@@ -34,12 +33,11 @@ const GraphQLVendorInfo = new GraphQLObjectType({
 export const GraphQLVendor = new GraphQLObjectType({
   name: 'Vendor',
   description: 'This represents a Vendor account',
-  interfaces: () => [GraphQLAccount, GraphQLAccountWithHost, GraphQLAccountWithContributions],
+  interfaces: () => [GraphQLAccount, GraphQLAccountWithContributions],
   isTypeOf: collective => collective.type === CollectiveType.VENDOR,
   fields: () => {
     return {
       ...AccountFields,
-      ...AccountWithHostFields,
       ...AccountWithContributionsFields,
       vendorInfo: {
         type: GraphQLVendorInfo,
